Change width/height to int from double

+++ src/main/java/com/grantranda/retorn/app/state/DisplayState.java
@@ -2,8 +2,8 @@ package com.grantranda.retorn.app.state;
 
 public class DisplayState {
 
-    private double width;
-    private double height;
+    private int width;
+    private int height;
     private boolean vSync;
 
     public DisplayState() {
